split out the main class into 3 logical steps
this pattern makes the module's logical steps easier to understand, and
abstracts away many ordering statements (which are represented already
in the top-level ordering)

This move also uses the opportunity to move to epp templates: we needed
to refactor all of the templates, anyway, in order to get values from
parameters of the main class. epp templates make it easier to see what
variables are used in templates and from where they come.