move service extra options to template file
the use of the inline_epp makes the manifest unreadable. arguably the
epp is not much more readable but it's not pushed out to its own file
and the puppet manifest should be cleaner.
This move also lets us use "puppet epp validate" more easily to make
sure that a syntax error like the one that was fixed in the previous
commit won't happen again.