start managing init.pre
the init.pre file contains some module load instructions that are
important, so this module should manage that file.

For now, we have no parameter that can change this file so we deploy it
as is. We'll use an epp template if we ever need to customize some parts
of the file.