Listen to both to IPv6 and v4 by default, and permit multiple addresses
The current default value only listens to IPv4, which might generate
useless connection errors in logs when spamc tries to connect with IPv6
first.

Also, the current data type for spamd_listen_address only permits
configuring one address to listen to where spamd can have more than one.
Moreover, the data type alias currently only permits IP addresses when
spamd can take hostnames, so the type is too strict.

By using an array of Stdlib::Host values, we fix both problems, but we
break the API, so this change means we will need to release a new major
version.

Finally, since we're changing the default to a value that will make
spamd listen to localhost both on IPv6 and IPv4, we also need to allow
connections from IPv6 as well.
Arguably, it was already possible to jam both allowed IPs in the same
string by separating addressed with commas, but I think the module's
interface is a lot easier to understand if we use an array for
spamd_allowed_ips: it's way more explicit that it supports more than one
entry, and it makes it easier for puppet user to give it multiple
values.