Fix default service flags for RHEL8
While trying to standup a new CentOS 8 system using this module I
discovered that the default service options were causing spamd to
constantly restart. After some in depth testing I found that the '-d'
option which may be needed for RHEL6 was causing the problem. As I no
longer have any RHEL7 based systems to test I don't know if it's needed
for version 7 or not, but I'm going to go out and assume that it is
since no updates were provided previously!

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>