Integrate new "cops" and enable them
This will silence the rubocop warning about their missing configuration.

Currently they are not triggering any issue so I'll just enable them
all. We can disable some later if they become a nuisance. I'm sure it's
not the Right Way (TM) to do this but I'm not a ruby programmer so I'll
approach this by forcing friction to see what works and not :P