Drop puppet 5.5 support
That release has been EOL for a while now and the CI tests currently
fail installing the puppet tools for running the tests. There's no good
reason to continue testing against this release.

The removal of support doesn't necessarily mean that the module won't
run with puppet 5.5 (there's a good chance that it'll continue running
just fine in the near future), but it means that the module is not
getting tested for it and workaround concerning that particular version
will not be accepted here.