Drop support for CentOS 6
CentOS 6 is EOL so this module should stop supporting it. CentOS 7 is
still getting maintenance updates until 2024 so we'll keep support for
it until the very end. And as mentioned in the readme, version 8 will be
there as long as version 7 is supported even though folks should
definitely migrate away from that version (but it is not this module's
mission to push ppl away from it)