Bump version to 3.2.0
This is strictly a release that bumps some ruby lib dependencies and,
most importantly, drops support for a couple OS and puppet releases (and
marks a new ubuntu release as supported).

So the module's code doesn't really change. The minor version bump is
mostly to signify the drop in support for the OS and puppet releases.