Fix spec tests
All of the puppet code and ruby code linting caused a bunch of problems
in the spec tests. I wasn't super attentive with the change of regex
delimiter from /../ to %r{..} and some special characters needed to be
escaped (I wonder if it was an issue beforehand already?)

I had also introduced changes in puppet code without changing spec tests
(bad LeLutin).

The tests were not defining the fact for the os major release so there
was no way they were running correctly before.

NOTE: I've decided to add major release 8 in the choices for RedHat
family in params.pp -- hopefully the default options are right for that
release, too.