CI weirdness config
the Gemfile uses the env var PUPPET_GEM_VERSION so I believe before this
change CI was not installing the specified version of puppet.

There's no system_tests group in the Gemfile.

Lastly the task to run all tests is currently using the default task but
I really just want that to run spec tests specifically.