Let gem determine appropriate version for puppetlabs_spec_helper
gem is a tool built specifically for resolving versions within
dependency constraints. We don't need to manually chose which version of
the lib is going to be used.

This simplifies things and it makes bundle install work when
PUPPET_GEM_VERSION is not specified -- this would error out because of a
conflict between voxpupuli-test and puppetlabs_spec_helper.