Bump puppet requirement to include 8.x
Somehow specifying only ">=6.0" was not prompting puppet_metadata to
include puppet 8 in the test matrix, but bounding the version up to 9.0
does include puppet 8 in the matrix o_O. sounds like a puppet_metadata
bug.

Anyways puppet 9.x does not exist yet and we still don't know if this
module will work properly with it.

This change will make CI run tests against puppet 8.x so we can start
claiming compatibility with it.