Fix rubocop errors discovered by new voxpupuli-test
renovate doesn't have a good understanding of those gem specifications
fenced by ruby version. it had changed both version of voxpupuli-test to
the same version. I think enough time has passed and we can try to
reduce this line to just one spec for the v7.0.0 of the gem and drop
compat code for ruby before 2.7.0