Retire support for puppet 6. Add ubuntu 24.04 to support list
puppet 6.x is by now EOL and to keep up to date with json gem
dependency, the ruby version that came with puppet 6 is no longer
working.

The module will probably keep working with puppet 6.x for a while, but
might break at some point since it won't be tested anymore.