Doc: Remove duplicate useless default values. reorder,group
The default values are already extracted by puppet-strings so it's
redundant to specify them in the top of file documentation.

Some of the mentioned default values were not really default values for
the parameters but rather the upstream spamd default value.

To make the documentation more compact, remove empty lines between
params, but make sure they are grouped similarly to the params on the
class so that it's still readable.

Some of the default values were plain wrong (or maybe very ancient
values?)

Update upstream reference URLs to https and version 4.0.x. Some of the
URLs were not correctly pointing to the right place in the page, so fix
the selectors.

Also add a URL to all of spamassassin's upstream documentation to be
more helpful to users.