readme: remove reference section. fix versions in limitations
Now that we have a REFERENCE.md file, including this information in the
README is redundant and the file is a lot harder to read because of that
section. Also since the information in the README is updated manually,
it tends to bitrot and be inaccurate or incomplete. Users should now
instead refer to the REFERENCE file.

The versions mentioned in the limitations section were not updated
before tagging 3.3.0. woops.. now the info is more up to date.

There's one old URL to upstream documentation. Point ppl to SA's 4.0.x
documentation over https instead.