Remove most gem upper bounds
Setting random upper bounds is super counter-productive. We'll set one
if ever there's a breaking change that's detected.

Also with upper bounds, I get poked ever now and then by renovate bot to
bump the version of a dependency. It's annoying for a module that sees
very infrequent maintenance.

I guess the upshot of having interactions with renovate bot is that the
PRs that it creates provokes running a test with the new versions. We'll
be losing that. For now we can live without this.